+++ src/Estonia.cpp
@@ -43,7 +43,7 @@ void Estonia::processChannel(int c, Input& in, Output& out) {
 		
 	simd::float_4 v = in.getPolyVoltageSimd<simd::float_4>(c);
 	audioFilter[c/4].setParameters(afp);
-	out.setVoltageSimd(simd::clamp(audioFilter[c/4].processAudioSample(v),-5.f,5.f),c);
+	out.setVoltageSimd(simd::clamp(audioFilter[c/4].processAudioSample(v),-10.f,10.f),c);
 }